backend/requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
spacy
scispacy
requests
neo4j
networkx
python-dotenv